Add maxRows parser method

+++ module/ldbc-generator/src/main/scala/ldbc/generator/model/Table.scala
@@ -26,7 +26,7 @@ object Table:
     case class EngineAttribute(value: String)               extends Options
     case class InsertMethod(value: "NO" | "FIRST" | "LAST") extends Options
     case class KeyBlockSize(value: Int)                     extends Options
-    case class MaxRows(value: Int)                          extends Options
+    case class MaxRows(value: Long)                          extends Options
     case class MinRows(value: Int)                          extends Options
     case class PackKeys(value: "0" | "1" | "DEFAULT")       extends Options
     case class Password(value: String)                      extends Options
